Extracted helper functions to a separate file
